add tests files
